RIP-458, split canvas_utils.py into separate site and user utils
