+++ ripley/api/canvas_site_controller.py
@@ -31,8 +31,9 @@ from ripley.externals import canvas, data_loch
 from ripley.externals.redis import enqueue, get_job
 from ripley.lib.berkeley_course import sort_course_sections
 from ripley.lib.berkeley_term import BerkeleyTerm
-from ripley.lib.canvas_utils import canvas_section_to_api_json, canvas_site_to_api_json, create_canvas_project_site, \
-    get_official_sections, get_teaching_terms, provision_course_site, update_canvas_sections
+from ripley.lib.canvas_site_utils import canvas_section_to_api_json, canvas_site_to_api_json, \
+    create_canvas_project_site, get_official_sections, get_teaching_terms, provision_course_site, \
+    update_canvas_sections
 from ripley.lib.http import tolerant_jsonify
 from ripley.lib.util import to_bool_or_none
 from ripley.merged.grade_distributions import get_grade_distribution_with_demographics, \
