+++ ripley/jobs/export_term_enrollments_job.py
@@ -32,7 +32,7 @@ from ripley.externals.s3 import upload_dated_csv
 from ripley.jobs.base_job import BaseJob
 from ripley.jobs.errors import BackgroundJobError
 from ripley.lib.berkeley_term import BerkeleyTerm
-from ripley.lib.canvas_utils import format_term_enrollments_export
+from ripley.lib.canvas_site_utils import format_term_enrollments_export
 from ripley.lib.util import utc_now
 from ripley.models.canvas_synchronization import CanvasSynchronization
